refactor: move useCanvasInteractions to follow DDD structure
Move useCanvasInteractions from src/composables/graph/ to src/renderer/core/canvas/
to follow Domain-Driven Design principles. Canvas interactions are specifically
about canvas behavior and should be grouped with other canvas-related functionality.

Also move the corresponding test file to maintain the same directory structure.

Updated all import paths across:
- Vue components (GraphCanvas, GraphCanvasMenu, SelectionToolbox, etc.)
- Vue node components (LGraphNode, NodeWidgets)
- Composables (useNodeEventHandlers, useNodeImage)
- Test files and mocks
